% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnbsr-package.R
\docType{package}
\name{cnbsr-package}
\alias{cnbsr}
\alias{cnbsr-package}
\title{cnbsr: acesso ao Catálogo Nacional de Bens e Serviços}
\description{
O pacote \code{cnbsr} fornecerá ferramentas para consultar o Catálogo Nacional
de Bens e Serviços (CNBS). No MVP, o escopo é limitado a Materiais.
}
\author{
\strong{Maintainer}: Raul Durlo \email{durlo.raul@gmail.com}

Authors:
\itemize{
  \item Raul Durlo \email{durlo.raul@gmail.com}
}

}
\keyword{internal}
