% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dados-basicos-pdm-codigo.R
\name{get_dados_basicos_pdm_por_codigo}
\alias{get_dados_basicos_pdm_por_codigo}
\title{Recuperar dados básicos de PDM por código}
\usage{
get_dados_basicos_pdm_por_codigo(codigo_pdm)
}
\arguments{
\item{codigo_pdm}{Número inteiro positivo correspondente ao código do PDM.}
}
\value{
Um \code{tibble} com uma linha e as nove colunas retornadas pela API.
Uma consulta válida sem resultado retorna zero linhas e as mesmas nove
colunas.
}
\description{
Consulta o endpoint \code{dadosbasicospdmporcodigo} do catálogo de Materiais do
CNBS para recuperar os dados básicos e a hierarquia de um PDM.
}
\examples{
\dontrun{
get_dados_basicos_pdm_por_codigo(17708)
}

}
