% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dados-item-material-codigo.R
\name{get_dados_item_material_por_codigo}
\alias{get_dados_item_material_por_codigo}
\title{Recuperar a descrição de um item de material por código}
\usage{
get_dados_item_material_por_codigo(codigo_item_material)
}
\arguments{
\item{codigo_item_material}{Número inteiro positivo correspondente ao código
do item de material.}
}
\value{
Um \code{tibble} com uma linha e a coluna \code{descricaoItem}. Uma consulta
válida sem resultado retorna zero linhas e a mesma coluna.
}
\description{
Consulta o endpoint \code{dadosItemMaterialporCodigo} do catálogo de Materiais
do CNBS para recuperar a descrição completa de um item.
}
\examples{
\dontrun{
get_dados_item_material_por_codigo(267203)
}

}
