% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/material-caracteristica-valor-pdm-sem-filtro.R
\name{get_material_caracteristica_valor_pdm_sem_filtro}
\alias{get_material_caracteristica_valor_pdm_sem_filtro}
\title{Recuperar itens, características e valores por PDM sem filtro}
\usage{
get_material_caracteristica_valor_pdm_sem_filtro(codigo_pdm)
}
\arguments{
\item{codigo_pdm}{Número inteiro positivo correspondente ao código do PDM.}
}
\value{
Um \code{tibble} com o mesmo contrato documentado em
\code{\link[=get_material_caracteristica_valor_pdm]{get_material_caracteristica_valor_pdm()}}.
}
\description{
Função mantida por compatibilidade com a versão 0.1.0. Equivale a chamar
\code{\link[=get_material_caracteristica_valor_pdm]{get_material_caracteristica_valor_pdm()}} com \code{com_filtro = FALSE}.
}
\examples{
\dontrun{
get_material_caracteristica_valor_pdm_sem_filtro(348)
}

}
