% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unidade-fornecimento-item-material.R
\name{get_unidade_fornecimento_por_codigo_item_material}
\alias{get_unidade_fornecimento_por_codigo_item_material}
\title{Recuperar unidades de fornecimento de um item de material}
\usage{
get_unidade_fornecimento_por_codigo_item_material(codigo_item_material)
}
\arguments{
\item{codigo_item_material}{Número inteiro positivo correspondente ao código
do item de material.}
}
\value{
Um \code{tibble} com uma linha por unidade de fornecimento e as colunas
\code{siglaUnidadeFornecimento}, \code{nomeUnidadeFornecimento},
\code{capacidadeUnidadeMedida}, \code{siglaUnidadeMedida} e \code{nomeUnidadeMedida}. Uma
consulta válida sem resultados retorna zero linhas e as mesmas colunas.
}
\description{
Consulta o endpoint \code{unidadeFornecimentoPorCodigoItemMaterial} do catálogo
de Materiais do CNBS para recuperar as unidades de fornecimento associadas a
um item.
}
\examples{
\dontrun{
get_unidade_fornecimento_por_codigo_item_material(267203)
}

}
