% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnbsr-package.R
\docType{package}
\name{cnbsr-package}
\alias{cnbsr}
\alias{cnbsr-package}
\title{cnbsr: acesso ao Catálogo Nacional de Bens e Serviços}
\description{
O pacote \code{cnbsr} fornece ferramentas para consultar o Catálogo Nacional de
Bens e Serviços (CNBS). Na versão 0.1.1, o escopo é limitado a Materiais,
com busca textual, consultas por código e navegação pela hierarquia do
catálogo.
}
\author{
\strong{Maintainer}: Raul Durlo \email{durlo.raul@gmail.com}

Authors:
\itemize{
  \item Raul Durlo \email{durlo.raul@gmail.com}
}

}
\keyword{internal}
